TABLE code_authority 추가

+++ src/main/java/com/github/bestheroz/demo/api/entity/member/MemberEntity.java
@@ -30,7 +30,7 @@ public class MemberEntity extends AbstractCreatedUpdateEntity implements Seriali
 
   private String password;
   private String name;
-  private Integer authorityId;
+  private Long authorityId;
   private Integer loginFailCnt;
   private Boolean available;
   private String theme;
